Fixed majot bug in computing dohtbydoct
In this commit, I have fixed a major bug in computing the gradient of ht wrt to ot. dohtbydoot should have been tanh(ct) but
I had wrongly put it as derivative of tanh(ct). I identified the bug while carefully going through the code
one function at a time!
The following are the other changes made in this commit:
1. Changed the initialization of ita for each of the gates. I was previously just copying the contents of outputMatrix. Now I'm explicitly initilaizing them as all zeros.
2. Passing the trainOrTestString as an argument to the forwardpass_lstm function. This is then passed as argument to the MLFFNN class. This is required for Batch Normalization in FC layer.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>