Fixed bug in computing previous cell state
In this commit, I have fixed a bug in computing Ct-1 when t=0. I was sampling the cellStateCurrent[t-1,:,:] to obatin the previous cell state. This is fine as long as t >=1. But when t is 0, then it becomes
cellStateCurrent[-1,:,:]! Which means it will sample the last/latest cell state post the forward pass! Where as I wanted the cell state at t=-1. SO this was a bug which was identified by chatGPT 5.0 version. This was a very good catch. I have now fixed it by sampling cellStatePrevious[t,:,:] instead of cellStateCurrent[t-1,:,:]. This bug was present in Forget gate contribution of gradient_ht_from_top.

The other major issue I was facing was the stack crashing due to very large value input to the tanh activation function. On caeful analysis, I observed that the cell state was drfitng to very large values across epochs. Also, when I montored the do L / do ct and do L/ do ht, it was hitting 60,000! All this was causing the code to crash across 1500 epochs with a LR = 1e-2. After a lot of discussion with chatGPT, it suggested to keep the gradients low by normalizing across time steps as well. And so, I have replaced the sum of gradients across time steps with mean across time steps to lower the weight update and prevent large values at the input of the tanh activation and thus prevent crashes. Now, I dont see any code crashes even after 2000 epochs even with a hig lr of 1e-2. Tis is because, we have further scaled down the gradients by the time steps.

LSTMs (and in fact RNNs) suffer from exploding gradients problem. Due to this, the update to the parameter weights becomes large and this can cause stack crashes when evaluating the exp of very large values. To avoid this, I have implemented a global norm clipping which normalizes the gradient values so that the updates to the parameter weights doesnt blow up and is controlled. I have removed the element wise clipping of the gradients since this is not the correct way to do and causes non differentiable operations in the gradient flow. In global norm clipping, we first compute the L2 norm squared of each of the parameter gradients (across time and layers) and then sum them all up and then take square root to obtain a global L2 normal of all the gradient parameters across time and layers. If the global norm is above some threshold say maxNorm, then we scale all the gradients to cap the norm to maxNorm. If the global norm is within the maxNorm margin, then we dont do anything. This is called the global norm clipping. For this, I have introduced some list variables to store all the gradients across time and layers.

Following are the other changes made in this commit:
1. While computing the loss function for categorical cross entropy, I was masking predicted values which were 0s to avoid them in the log computation. But this is not correct, since this will mask cases where the true label is 1. This is not the correct way to do. Instead just add a small eps to avoid 0s in the log computation. I have mde this change both in the LSTM as well as DNN class.
2. Added some commented out print statements to monitor the norm of do L / do ct, do L / do ht, max, min, mean values of the cell states, etc. These are debug hooks to check for exploding gradients and cell state run off.
3. Broke up the update_weights function into 2 sub functions, namely compute_gradients(which only computes the gradients and also the global norm) followed by update_weights(which only does weight update using Gradient descent).

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>