Fix referencing bug
In this commit, I have fixed a major bug in initializing the gates as a
list. I was generating an array of zeros and appending this same array
to each of the gates. This is incorrect. SInce it is the same array that
is getting appended to each of the gates, each array in each of the
gates points to the same array. So, changing any one of the gates
entries will change all the others since they are all referencing the
same array. This was major bug I identifed while bitmatching with
pytorch based implementation. I have fixed this now. However, the loss
hasn't improved! There seems to be a bug somewhere else also. The
gradients of ht and ct are matching with the pytorch implementation for
the last 2 time steps only. From T-2 to 1 time steps, it is not
matching! I have spent 2 full days on this and couldn't identify the
issue.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>