Implement vanilla GAN
In this commit, I have implemented a vanilla GAN on MNIST dataset. GAN stands for Generative Adverserial Networks.
I have used the architecture as mentioned in the tutorials of Deep Generative Models by Pratosh.
Link:
    https://www.youtube.com/watch?v=iOb8vmlJd8o&list=PLZ2ps__7DhBa5xCmncgH7kPqLqMBq7xlu&index=13
The implementation is in pytorch. The code is well documented to explain each and every step.
But the results are not so great because, I have done a simple flattening of the images and feeding into the
Discriminator network which is a Feed Forward neural network. Similarly, at the Generator, I'm using a FFNN and finally reshaping
the ouput to a 28 x 28 image. Ideally, I should be using convolution NNs to process images. I wil be doing this in my next
where I will be implementing a Conditional Deep Convolutional GAN. This implementation is to get started on GAN implementation
and get hands-on pytorch.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>