Compute Encoder and decoder gradients with one backward pass
Previously, I was implementing a separate weight updates for the encoder and decoder. While implementing the
Encoder weight update, I was using the entire ELBO(Evidence Lower Bound) cost function which comprises both the
reconstruction/ negative log likelihood term and the KL divergence between q_phi (z|x) and p(z). While updating
the Decoder parameters, I was considering only the reconstruction term and not the KL term since the KL term
depends only on the Encoder parameters. But in this commit, I have clubbed both the Encoder and Decoder
updates into a single forward and backward path and updates for both of them together. This is because,
autograd (pytorch's toll for computig gradients) forms a computational graph during the forward pass
and hence know the paths to follow during the backward pass to compute the gradients.
Hence, I can just give the total loss and autograd take care of computing the gradients for both the Encoder
and Decoder.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>