Generate Human faces using VAE
In this commit, I have implemented the VAE on human faces. The human faces data is available from a famous dataset called celebA dataset. The specific architecture was suggested to me by chatGPT
and it is based on architectures presented in several influential works and implementation.
The results are surprisingly good! It is able to generate decent human faces! With this, I have successfully
implemented both VAEs and GANs. Next, I will move to DDPMs which are supposedly state of the art.

Since the celebA dataset is huge, it cannot be downloaded using the datasets module of torchvision. Hence, I have separately downloaded the data and have written a helper script named 'celeba_dataset', which helps apply the transform function to the data.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>