Implement DDPM by regressing over mean
In this commit, I have added a script which implements a variant of DDPM which regresses over the mean of the known denoising distribution q. The implementation is based on the video lectures of Pratosh on Deep Generative Models. I have also derived the expression for the objective of the DDPM (which is a variant of ELBo for hierarchical latent spaces). Also, ideally, when training, the consistency /denoising mathcing term should be computed as a sum of losses for each minibatch summed over M number of time steps. This is based on the lectures of Pratosh. However, this would need looping over the subset of time steps and then summing the loss of each mini batch of examples. This is extremely memory and compute intensive. I will change this approach to what is actually done in the original DDPM paper. Since the current implementation is very memory intensive, when run on a heavy dataset like celebA, I was running into CUDA memory issues and the laptop was hanging eventually leading to code crash. So, affectively, I couldnt get to see any results whatsoever with this implementation of DDPM.  In the next commit, I will modify the time sampling and reduce the compute so that I can alteast run the code.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>