20/09/2025
Modified Time Sampling strategy for training

In this commit, I have modified the time sampling strategy used for training on a mini-batch. In the last commit,
I was computing the consistency/denoising matching term loss for a given batch of data samples for a subset of
Time steps. This meant I was computing the loss for a mini batch for a time index t and
then looping over the subsets of T and summing the losses over all the subsets of T. This was a memory and compute
heavy operation and I couldnt even get the code running. It was crashing due to 'out of memory on GPU' issues.
But this was the implementation as discussed in Pratosh video tutorials.
However, the time sampling is not done this way in the original paper on DDPM. For each example in the mini-batch,
they compute the consistency term loss for different time steps T.
In other words, each example has a different time step T over which the loss is computed. We then sum the losses
of each example of the mini-batch computed for its corresponding time step T. This method of computing
the consistency term loss is much more compute and memory friendly! I have implemented this method of time sampling
in this commit. So, I could get the code to run atleast! However, the results/generated images are not good at all!
The loss is behaving well and is reducing across epochs but this does not reflect in the quality of generated images.
Even in the original paper on DDPM, they don't regress over mu_q i.e the mean of the reverse denoising distribution
q(X_t-1|X_t, X_0). In fact, in the paper, they regress over the noise epsilon i.e nise added to X_0 to land at X_t.
In the nect commit, I will regress over the noise instead of on the mean and see if I can get meaningful generated
images. But I was still not very sure as to why the regression over mu was not yielding good generated images even
though the training loss was dropping. On checking online with Google Gemini, it said that regression over mu
doesnt yield good results for the following reason:
    When regressing over mean mu_q, the mu_q is a function of X_T and X_0. Hence the output of the U-Net which
    is mu_theta is regressing over mu_q which is a function of the input. Hence, during training, across epochs,
    the U-Net NN which is outputing mu_theta(mean of the learnable denoising distribution at time t,
    P_theta(X_t-1|X_T)) starts coming closer to mu_q(which is a fucntion of the input image). Hence, across epochs,
    the fit becomes better and the loss drops. So, the U-Net has learnt to condition on the input image. However,
    during generation, there is no input image X_0 to condition on and also the U-Net has not learnt to generate
    images without conditioning on the input image. hence, at inferece, it generates garbage.
    This is what is the explanation from Google Gemini!

However, in the equation for mu_q(which is a function of X_t and X_0), if we replace X_0 in terms of X_t and epsilon
(using the equation for formward noising process), now mu_q becomes a function of X_t and noise epsilon.
Now there is no dependence on X_0. So, mu_q is now a function of X_t and noise epsilon(noise added to X_0 to get to X_t).
With this formulation, the learnable mu_theta also becomes a function of X_t and predicted noise epsilon_theta.
With this formulation, the U-Net outputs epsilon_theta instead of mu_theta. Hence, the problem of regressing over means
mu becomes a problem of regression over epsilon! I will make this change in the next script where I will regress over noise!

Also, in the original paper, the reconstruction term loss is not included and hence I have commented out in the code
as well.

I have implemeted U-Net based on the recommendation by chat GPT. U-Net is a DNN which has the following architecture:
    1. Encoder layers
    2. BottleNeck Layer
    3. Decoder layers
    4. Skip connections or residual connections

The encoder layers increase the size of the feature maps but reduce the size of the image and the decoder
layers gradually decrease back the size of the feature maps while increasing back the size of the image.
U-Net is a popular DN architecture used in DDPMs, image segmentation etc. Also, In the DDPM implementation,
we also feed in the time step as an input to the U-Net. This is to condition the denoising learning on the time step
Hence, the network learns to denoise based on what time step it is operating at. In the original DDPM paper,
they added the time embedding to the U_Net and the time embedding was done using a sinusoidal positional embedding.
So, I have also implemented the sinusoidal time mebedding and added it at each layer of the encoder,
bottleneck layer, decoder.
Also, renamed the script from ddpm_celeba --> ddpm_celeba_regressmean

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>